% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_adjusted_forest.R
\name{plot.ggadjustedforest}
\alias{plot.ggadjustedforest}
\title{Plot method for ggadjustedforest objects}
\usage{
\method{plot}{ggadjustedforest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggadjustedforest}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the plot object.
}
\description{
Plot method for ggadjustedforest objects
}
