% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_multi_outcome_forest.R
\name{plot.ggadjustedforest_multi}
\alias{plot.ggadjustedforest_multi}
\title{Plot method for ggadjustedforest_multi objects}
\usage{
\method{plot}{ggadjustedforest_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggadjustedforest_multi}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the plot object.
}
\description{
Plot method for ggadjustedforest_multi objects
}
