% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_adjusted_forest.R
\name{print.ggadjustedforest}
\alias{print.ggadjustedforest}
\title{Print method for ggadjustedforest objects}
\usage{
\method{print}{ggadjustedforest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggadjustedforest}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for ggadjustedforest objects
}
