% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_multi_outcome_forest.R
\name{print.ggadjustedforest_multi}
\alias{print.ggadjustedforest_multi}
\title{Print method for ggadjustedforest_multi objects}
\usage{
\method{print}{ggadjustedforest_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggadjustedforest_multi}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for ggadjustedforest_multi objects
}
